move get_nearest_graphics to utils
