+++ fastplotlib/layouts/_plot_area.py
@@ -357,80 +357,6 @@ class PlotArea:
         # default z is zero for now
         return np.array([*pos_world[:2], 0])
 
-    def get_nearest_graphics(
-        self,
-        pos: tuple[float, float] | tuple[float, float, float],
-        method: str = "center",
-        subset: GraphicCollection | np.ndarray[Graphic] = None,
-    ) -> np.ndarray[Graphic]:
-        """
-        Returns the nearest ``n_graphics`` to the passed position ``pos`` in world space
-
-        Parameters
-        ----------
-        pos: (x, y) | (x, y, z)
-            position in world space, not data space
-
-        method: "center" | "edge"
-            Determine the nearest objects using the "center" of graphics or the "edge".
-            Note that the "edge" method uses an axis-aligned bounding box and therefore
-            maybe be inaccurate if the Graphic is not axis-aligned, ex. if it is rotated
-
-        subset: np.ndarray[Graphic], optional
-
-
-        Returns
-        -------
-        tuple[Graphic]
-            nearest graphics to ``pos`` in order
-
-        """
-
-        if subset is not None:
-            if isinstance(subset, GraphicCollection):
-                subset = subset.graphics
-
-            if not isinstance(subset, np.ndarray):
-                raise TypeError("`subset` must be a numpy array of Graphic objects")
-
-            if not all(isinstance(g, Graphic) for g in subset):
-                raise TypeError("all elements of `subset` must be Graphic objects")
-
-            graphics = subset
-
-        else:
-            graphics = self.graphics
-
-        pos = np.asarray(pos)
-
-        if pos.shape != (2,) or not pos.shape != (3,):
-            raise TypeError
-
-        match method:
-            case "center":
-                # use sphere
-                centers = np.empty(shape=(len(graphics), len(pos)))
-                for i in range(centers.shape[0]):
-                    centers[i] = graphics[i].world_object.get_world_bounding_sphere()[:len(pos)]
-
-                distances = np.linalg.norm(centers[:, :len(pos)] - pos, ord=2, axis=1)
-
-            case "edge":
-                # use bbox
-                distances = np.empty(shape=(len(graphics),))
-                for i in range(distances.size):
-                    bbox = graphics[i].world_object.get_world_bounding_box()
-
-                    distances[i] = np.linalg.norm(
-                        np.abs(bbox[:, :len(pos)] - pos).min(axis=0)
-                    )
-
-            case _:
-                raise ValueError("method must be one of 'center' or 'edge'")
-
-        sort_indices = np.argsort(distances)
-        return np.asarray(graphics)[sort_indices]
-
     def set_viewport_rect(self, *args):
         self.viewport.rect = self.get_rect()
