+++ fastplotlib/utils/__init__.py
@@ -3,7 +3,7 @@ from dataclasses import dataclass
 
 from .functions import *
 from .gpu import enumerate_adapters, select_adapter, print_wgpu_report
-
+from ._plot_helpers import *
 
 @dataclass
 class _Config:
