+++ fastplotlib/utils/_plot_helpers.py
@@ -1,0 +1,53 @@
+from typing import Sequence
+
+import numpy as np
+
+from ..graphics._base import Graphic
+from ..graphics._collection_base import GraphicCollection
+
+
+def get_nearest_graphics(
+        graphics: Sequence[Graphic] | GraphicCollection,
+        pos: tuple[float, float] | tuple[float, float, float],
+) -> np.ndarray[Graphic]:
+    """
+    Returns the nearest ``n_graphics`` to the passed position ``pos`` in world space.
+    Uses the distance between ``pos`` and the center of the bounding sphere for each graphic.
+
+    Parameters
+    ----------
+    graphics: Sequence, i.e. array, list, tuple, etc. of Graphic | GraphicCollection
+        the graphics from which to return a sorted array of graphics in order of closest
+        to furthest graphic
+
+    pos: (x, y) | (x, y, z)
+        position in world space, z-axis is ignored when calculating L2 norms if ``pos`` is 2D
+
+    Returns
+    -------
+    tuple[Graphic]
+        nearest graphics to ``pos`` in order
+
+    """
+
+    if isinstance(graphics, GraphicCollection):
+        graphics = graphics.graphics
+
+    if not all(isinstance(g, Graphic) for g in graphics):
+        raise TypeError("all elements of `graphics` must be Graphic objects")
+
+    pos = np.asarray(pos)
+
+    if pos.shape != (2,) or not pos.shape != (3,):
+        raise TypeError
+
+    # get centers
+    centers = np.empty(shape=(len(graphics), len(pos)))
+    for i in range(centers.shape[0]):
+        centers[i] = graphics[i].world_object.get_world_bounding_sphere()[:len(pos)]
+
+    # l2
+    distances = np.linalg.norm(centers[:, :len(pos)] - pos, ord=2, axis=1)
+
+    sort_indices = np.argsort(distances)
+    return np.asarray(graphics)[sort_indices]
